requirements.txt
absl-py==1.4.0
addict==2.4.0
asttokens==2.2.1
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///C:/b/abs_4a0polqwty/croot/certifi_1683875377622/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama==0.4.5
contourpy==1.1.0
cryptography @ file:///C:/b/abs_b28tzldafl/croot/cryptography_1686613103834/work
cycler==0.11.0
decorator==4.4.2
einops==0.4.1
executing==1.2.0
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
fonttools==4.41.1
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
huggingface-hub==0.16.4
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipython==8.12.2
IQA-pytorch==0.1
jedi==0.18.2
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
kiwisolver==1.4.4
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe @ file:///C:/ci/markupsafe_1654489871526/work
matplotlib==3.7.2
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.2.1
natsort==8.2.0
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
numpy==1.24.4
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.5.5.62
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
protobuf==4.23.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.15.1
pyiqa==0.1.7
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytorch-ssim==0.1
pytorch-wavelets==1.3.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.6.3
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tb-nightly==2.14.0a20230722
tensorboard-data-server==0.7.1
tifffile==2023.7.10
timm==0.9.2
tomli==2.0.1
torch==2.0.1
torch-summary==1.4.5
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions @ file:///C:/b/abs_5em9ekwz24/croot/typing_extensions_1686602003259/work
tzdata==2023.3
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
wcwidth==0.2.6
Werkzeug==2.3.6
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
yacs==0.1.8
yapf==0.40.1
zipp==3.16.2
